Fix slice bounds out of range on address copy
An IP address embedded in a net.TCPAddr (or net.UDPAddr) struct returned
from net.ResolveTCPAddr (or net.ResolveUDPAddr) can be either a 4-byte
or 16-byte slice depending on platform and configuration. Therefore, we
must check the length of the address before copying it into the
syscall.Sockaddr to avoid a "slice bounds out of range" panic.